+++ src/main/java/com/iexec/core/task/update/TaskUpdateManager.java
@@ -65,8 +65,7 @@ class TaskUpdateManager {
     private final BlockchainAdapterService blockchainAdapterService;
     private final SmsService smsService;
 
-    private final Map<TaskStatus, AtomicLong> currentTaskStatusesCount;
-    private final ExecutorService taskStatusesCountExecutor;
+    private final LinkedHashMap<TaskStatus, AtomicLong> currentTaskStatusesCount;
 
     public TaskUpdateManager(TaskService taskService,
                              IexecHubService iexecHubService,
@@ -97,23 +96,31 @@ class TaskUpdateManager {
                             "status", status.name()
                     ).register(Metrics.globalRegistry);
         }
-
-        this.taskStatusesCountExecutor = Executors.newSingleThreadExecutor();
     }
 
     @PostConstruct
     Future<Void> init() {
-        return taskStatusesCountExecutor.submit(
-                // The following could take a bit of time, depending on how many tasks are in DB.
-                // It is expected to take ~1.7s for 1,000,000 tasks and to be linear (so, ~17s for 10,000,000 tasks).
-                // As we use AtomicLongs, the final count should be accurate - no race conditions to expect,
-                // even though new deals are detected during the count.
-                () -> currentTaskStatusesCount
-                        .entrySet()
-                        .parallelStream()
-                        .forEach(entry -> entry.getValue().addAndGet(taskService.countByCurrentStatus(entry.getKey()))),
+        final ExecutorService taskStatusesCountExecutor = Executors.newSingleThreadExecutor();
+        final Future<Void> future = taskStatusesCountExecutor.submit(
+                this::initializeCurrentTaskStatusesCount,
                 null    // Trick to get a `Future<Void>` instead of a `Future<?>`
         );
+        taskStatusesCountExecutor.shutdown();
+        return future;
+    }
+
+    /**
+     * The following could take a bit of time, depending on how many tasks are in DB.
+     * It is expected to take ~1.7s for 1,000,000 tasks and to be linear (so, ~17s for 10,000,000 tasks).
+     * As we use AtomicLongs, the final count should be accurate - no race conditions to expect,
+     * even though new deals are detected during the count.
+     */
+    private void initializeCurrentTaskStatusesCount() {
+        currentTaskStatusesCount
+                .entrySet()
+                .parallelStream()
+                .forEach(entry -> entry.getValue().addAndGet(taskService.countByCurrentStatus(entry.getKey())));
+        publishTaskStatusesCountUpdate();
     }
 
     void updateTask(String chainTaskId) {
@@ -710,10 +717,28 @@ class TaskUpdateManager {
     void updateMetricsAfterStatusUpdate(TaskStatus previousStatus, TaskStatus newStatus) {
         currentTaskStatusesCount.get(previousStatus).decrementAndGet();
         currentTaskStatusesCount.get(newStatus).incrementAndGet();
+        publishTaskStatusesCountUpdate();
     }
 
     @EventListener(TaskCreatedEvent.class)
     void onTaskCreatedEvent() {
         currentTaskStatusesCount.get(RECEIVED).incrementAndGet();
+        publishTaskStatusesCountUpdate();
+    }
+
+    private void publishTaskStatusesCountUpdate() {
+        // Copying the map here ensures the original values can't be updated from outside this class.
+        // As this data should be read only, no need for any atomic class.
+        final LinkedHashMap<TaskStatus, Long> currentTaskStatusesCountToPublish = currentTaskStatusesCount
+                .entrySet()
+                .stream()
+                .collect(Collectors.toMap(
+                        Map.Entry::getKey,
+                        entrySet -> entrySet.getValue().get(),
+                        (a, b) -> b,
+                        LinkedHashMap::new
+                ));
+        final TaskStatusesCountUpdatedEvent event = new TaskStatusesCountUpdatedEvent(currentTaskStatusesCountToPublish);
+        applicationEventPublisher.publishEvent(event);
     }
 }